# ocl kernel run performance improvement