don't set cl event callback after kernel enqueue

This is not efficient if there are many ocl kernels to be enqueued,
and will impact application performance for example the dnn module.
So the patch just cleanup umats and images after kernel is enqueued.