@alalek 2017-11-13
UMat can be allocated over user data.
1. Sometimes OpenCV can't use `CL_MEM_USE_HOST_PTR` feature to do this - it depends on platform restrictions (non-unified memory between, unsupported alignment, etc) or just not supported in OpenCL runtime at all. In this case OpenCV simulate this behaviour: on allocation - makes "copy" of user data (via `CL_MEM_COPY_HOST_PTR` or via separate "copy" process) and after deallocation synchronize data back. So deallocation process should be done very accurate - at least on the same OpenCL command queue (we don't use out-of-order queues yet).

2. OpenCV uses buffer pool to reduce number of allocations/deallocations. So "freed" UMat can be reused again while its kernel still processing.