@pengli 2017-11-13
Is there a proper way to handle this situation  ? we observed "clSetEventCallback" cost some time in executing dnn net model like googlenet and we want to minimize this overhead.