@alalek 2017-11-13
Looks like, there is some handling via `haveTempDstUMats` flag (but this force full synchronization).
Perhaps we need something like this for source UMats.