@pengli 2017-11-14
@alalek , I added "haveTempSrcUMats" flag to check if src UMats requires async deallocation.  If yes,  cl event callback will be set, otherwise just call cleanupUMats directly. please check the new patch.