@alalek 2017-11-14
I'm observing 40% performance degradation on this test case (48 -> 79ms): `ENet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)`.
Could you check this on your platforms?