@pengli 2017-11-15
I run the test on  both Skylake (48eu) and Broadwell (24eu) ,  

1) on Skylake

without this PR:

[ PERFSTAT ]    (samples = 100, mean = 63.58, median = 63.25, stddev = 2.48 (3.9%))
[       OK ] DNNTestNetwork.ENet/1 (20120 ms)
[----------] 1 test from DNNTestNetwork (20120 ms total)

with this PR:

[ PERFSTAT ]    (samples = 100, mean = 71.73, median = 68.78, stddev = 7.82 (10.9%))
[       OK ] DNNTestNetwork.ENet/1 (19297 ms)
[----------] 1 test from DNNTestNetwork (19297 ms total)

2)  on Broadwell

without this PR:

[ PERFSTAT ]    (samples = 10, mean = 123.05, median = 121.77, stddev = 3.60 (2.9%))
[       OK ] DNNTestNetwork.ENet/1 (60136 ms)
[----------] 1 test from DNNTestNetwork (60136 ms total)

with this PR:

[ PERFSTAT ]    (samples = 38, mean = 116.94, median = 115.76, stddev = 3.29 (2.8%))
[       OK ] DNNTestNetwork.ENet/1 (57252 ms)
[----------] 1 test from DNNTestNetwork (57252 ms total)

So on skylake, the performance is down about 12%, on broadwell, the performance is improved.  one thing is that the torch enet model isn't fully enabled on GPU, there are many layers fallback called.  So I think it is better to check the OCL backend performance again when torch enet is fully enabled on GPU.