@pengli 2017-11-15
I also test googlenet on skylake which is now fully enabled on OCL path 

without this PR

[ PERFSTAT ]    (samples = 10, mean = 24.93, median = 24.78, stddev = 0.43 (1.7%))
[       OK ] DNNTestNetwork.GoogLeNet/1 (12785 ms)
[----------] 1 test from DNNTestNetwork (12785 ms total)

with this PR

[ PERFSTAT ]    (samples = 10, mean = 20.72, median = 20.62, stddev = 0.46 (2.2%))
[       OK ] DNNTestNetwork.GoogLeNet/1 (12531 ms)
[----------] 1 test from DNNTestNetwork (12531 ms total)

This PR clearly improve the performance.