@alalek 2017-11-20
Could you try this patch on your devices: https://github.com/alalek/opencv/commit/pr10076_r ?